made a new function indexTo15UniqueColor and used it instead of having duplicate code
