+++ app/src/miscUtil.ts
@@ -14,3 +14,25 @@ export function stringHash(input: string) {
 
 	return hash;
 }
+
+const memberColorArray: string[] = [
+	"bg-trip_member_col_1",
+	"bg-trip_member_col_2",
+	"bg-trip_member_col_3",
+	"bg-trip_member_col_4",
+	"bg-trip_member_col_5",
+	"bg-trip_member_col_6",
+	"bg-trip_member_col_7",
+	"bg-trip_member_col_8",
+	"bg-trip_member_col_9",
+	"bg-trip_member_col_10",
+	"bg-trip_member_col_11",
+	"bg-trip_member_col_12",
+	"bg-trip_member_col_13",
+	"bg-trip_member_col_14",
+	"bg-trip_member_col_15"
+];
+
+export function indexTo15UniqueColor(index: number) {
+	return memberColorArray[index % 15]
+}
